콘솔 기능 수정
